refactor(cmd): update cnosdb_inspect to cnosdb-inspect
