+++ cmd/cnosdb-inspect/deletetsm/deletetsm.go
