+++ cmd/cnosdb-inspect/verify/seriesfile/command.go
