+++ cmd/cnosdb-inspect/verify/seriesfile/verify.go
